% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegFlechsig-package.R
\docType{package}
\name{ggsegFlechsig-package}
\alias{ggsegFlechsig}
\alias{ggsegFlechsig-package}
\title{ggsegFlechsig: Flechsig Myelogenetic Atlas for the 'ggseg' Ecosystem}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Flechsig myelogenetic cortical atlas for the 'ggseg' ecosystem. Provides a unified 'ggseg_atlas' object with both 2D polygon geometry and 3D vertex indices, for use with 'ggseg' and 'ggseg3d'. Based on the supplementary materials of Pijnenburg et al. (2021).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegFlechsig}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegFlechsig/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

}
\keyword{internal}
